+++ PartitionDKaMinPar.h
